% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{objective}
\alias{objective}
\title{Define an objective function to optimize.}
\usage{
objective(fun)
}
\arguments{
\item{fun}{Function with an arbitrary number of parameters. Parameters with default
values don't need to be optimized, but they can if they're added as \code{\link[=variable]{variable()}}s.}
}
\value{
A \code{sugarmax_problem} object.
}
\description{
Define an objective function to optimize.
}
