% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sugarmax-package.R
\docType{package}
\name{sugarmax-package}
\alias{sugarmax}
\alias{sugarmax-package}
\title{sugarmax: Simple syntax for comfortable non-linear optimization.}
\description{
Lightweight wrapper around `stats::optim()` that allows multiple bounded multidimensional variables.
}
\author{
\strong{Maintainer}: benet \email{benet1one@gmail.com}

}
\keyword{internal}
