% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{variable}
\alias{variable}
\title{Add a variable to optimize.}
\usage{
variable(
  .problem,
  symbol,
  dimensions = 1,
  lower = -Inf,
  upper = +Inf,
  inits = NA
)
}
\arguments{
\item{.problem}{Problem created with \code{\link[=objective]{objective()}}.}

\item{symbol}{Name of the parameter in the objective function. Does not need to be a string.}

\item{dimensions}{Integer vector of length one or more giving the maximal indices in each dimension.
For instance, a scalar has a dimension of \code{1}, a vector of N elements has a dimension of \code{N},
and an NxM matrix has a dimension of \code{c(N, M)}.}

\item{lower}{Lower bound for the variable. Can be a scalar or an array of \code{dim = dimensions}.}

\item{upper}{Upper bound for the variable. Can be a scalar or an array of \code{dim = dimensions}.}

\item{inits}{Initial value for the variable. Can be a scalar or an array of \code{dim = dimensions}.}
}
\value{
A \code{sugarmax_problem} object.
}
\description{
Parameters in the objective function that aren't added will be left to their default value.
}
