% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{minimize}
\alias{minimize}
\alias{maximize}
\title{Optimize a problem.
Minimize or maximize a problem.}
\usage{
minimize(
  .problem,
  rel_tol = sqrt(.Machine$double.eps),
  abs_tol = NULL,
  method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"),
  ...
)

maximize(
  .problem,
  rel_tol = sqrt(.Machine$double.eps),
  abs_tol = NULL,
  method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent"),
  ...
)
}
\arguments{
\item{.problem}{Problem created with \code{\link[=objective]{objective()}} and \code{\link[=variable]{variable()}}.}

\item{rel_tol}{Relative convergence tolerance. The algorithm stops if it is unable to improve the
objective value by a factor of \code{reltol * (abs(val) + reltol)} at a step.}

\item{abs_tol}{The absolute convergence tolerance.
The algorithm stops if the objective value is better than \code{abs_tol}.}

\item{method}{The method to be used. See details in \code{\link[base:optim]{base::optim()}}.}

\item{...}{Other arguments passed to \code{\link[base:optim]{base::optim()}}.}
}
\value{
A list with elements:
\itemize{
\item \verb{$internal}: The output of \code{\link[base:optim]{base::optim()}}.
\item \verb{$objective_value}: Numeric, the value of the function at the optimum.
\item \verb{$solution}: Named list of numbers, the value of each variable at the optimum.
}
}
\description{
Optimize a problem.
Minimize or maximize a problem.
}
